format SIN display
