+++ sources/packages/web/src/composables/useFormatters.ts
@@ -8,7 +8,6 @@ import dayjs, { QUnitType, OpUnitType } from "dayjs";
 
 const DEFAULT_EMPTY_VALUE = "-";
 export const DATE_ONLY_ISO_FORMAT = "YYYY-MM-DD";
-const SIN_MAX_LENGTH = 9;
 
 /**
  * Helpers to adjust how values are shown in the UI.
@@ -183,29 +182,6 @@ export function useFormatters() {
     );
   };
 
-  const isSINValid = (sin: string): boolean => {
-    let valid = false;
-    if (sin) {
-      sin = sin.replace(/\s/g, "");
-      if (sin.length === SIN_MAX_LENGTH) {
-        let checksum = 0;
-        for (let i = 0; i < sin.length; i++) {
-          const currentDigit = +sin.charAt(i);
-          if ((i + 1) % 2 === 0) {
-            const digitTimes2 = currentDigit * 2;
-            checksum += digitTimes2 < 10 ? digitTimes2 : digitTimes2 - 9;
-          } else {
-            checksum += +sin.charAt(i);
-          }
-        }
-        if (checksum % 10 === 0) {
-          valid = true;
-        }
-      }
-    }
-    return valid;
-  };
-
   return {
     dateString,
     dateOnlyLongString,
@@ -217,6 +193,5 @@ export function useFormatters() {
     booleanToYesNo,
     sinDisplayFormat,
     getISODateOnlyString,
-    isSINValid,
   };
 }
